Removed ``Screen.default_char_dict``
Instead ``Screen.select_graphic_rendition`` now has a fast path for
resetting all attributes to their default values.

This commit has a behaviour change. Given an SGR sequence with 0 inside
``Screen`` would now ignore the 0 instead of resetting attributes in the
middle of the processing. Sadly ECMA-48 does not specify the expected
behaviour in that case, but the new implementation makes more sense to me.