Fix expected history positions in tests
The previous commit modifed the history position to track the index of
the last line of the screen in the history. This broke the tests, which
expected the position to move by the number of lines in the screen,
regardless of the actual fraction of the page being scrolled. This has
been corrected in this commit.