Expand history to full required length
A test is failing because the code is unable to reach the start of the
history, because the top buffer is too short, and the test code gets
stuck in an infinite loop. This change ensures the top buffer is long
enough to hold the whole history and that the test does not get stuck.