Update index test
Part c of the index test filled the history and checked that the top
buffer was of the correct length. With the history being expanded, the
previous strategy of filling by the square of the number of lines in
the terminal no longer works. Instead, we fill by twice the size of the
history.

Since the maximum size of the top buffer has been doubled, the assertion
for its length has also been updated here.