Allow temporal cursor_x > columns (x1.05 to x1.14 faster)
Instead of checking if cursor_x > columns at the end of iteration
and set cursor_x to the minimum of (cursor_x and columns), delay
that decision to the begin of the next iteration or at the end
of the for-loop.

This removes one "if" statement at the end of the for-loop and allows us
to use the local variable cursor_x all the time without having to update
cursor.x. Only this happens before insert_characters() and at the end of
the draw() method when the cursor.x is been visible by code outside
draw() and therefore must be updated with the latest value of cursor_x.

This optimization makes stream.feed between x1.05 and x1.14 faster. As
in any optimization on draw(), the use cases that gets more improvements
are the ones that have very few control sequences in their input (so
stream.feed is dominated by screen.draw and not be stream._parse_fsm)