Make Char mutable (not a namedtuple) (API may break)
Make Char mutable (and ordinary object) so we can modify each char in
place avoiding calling _replace.

This commit only changed the Char class and implements some methods to
emulate the namedtuple API.

Theoretically it could be possible to emulate the whole namedtuple API
but it is unclear if it worth. In this scenario, user code may break.

Using a plain object instead of a namedtuple added a regression on
memory usage of x1.20 for htop and mc benchmark files when HistoryScreen was
used. The rest of the benchmarks didn't change significantly (but it is
expected to be slightly more inefficient).