Update Chars in-place instead of recreating them (x1.20 and x1.90 faster; regress on mem)
Instead of calling _replace() to create a new Char object, modify the
existing one.

For that, the Line (ex StaticDefaultDict) is in charge to
fetch the char and do the modifications. If no Char is found, only then
a Char is created and inserted in the Line (dict). See write_data().

In some cases we need to get a Char, read it and then update it so a
copy of the Line's default char is returned and added to the line. A
copy is required because now the Char are mutable. See char_at().

Changed the API of Char: _asdict renamed as as_dict and  _replace as
copy_and_change; removed _replace_data and added a copy method.

The constructor also changed: it is required data, width and style. The
former way to construct a Char can be done with from_attributes class
method.

This commits improved the runtime of stream.feed by x1.20 to x1.90
(faster) however a regression on the memory footprint was found (between
x1.10 and x1.50). I don't have an explanation for this last point.