Fix a bug on index that top line was not removed.
If the next line of the margin's top was not empty, the for-loop used
that entry to override the top line working as expected.

But when the next line of the top was empty, the top line was untouched
so an explicit pop is required.

A similar issue happen on reverse_index.

Both bugs were covered due a side effect to iterating over
screen.buffer: on each line lookup, if no such exist, a new line is
added to the buffer. This added new line then was used to override the
top on an index() call.