Try to delete entries on erase instead of write spaces
If the line's default char and the cursor's attributes are the same,
instead of writing spaces into the line, delete the chars.

This should be equivalent from user's perspective.

This applies to erase_characters, erase_in_line and erase_in_display.

This optimization reduced the number of false lines (lines without any
char) and the number of blank lines (lines with only spaces).
With less entries in the buffer, the rest of the iterations can
take advantage of the sparsity of the buffer and process much less.