Make lines in history a LineView and make Line raise on non-key
Instead of storing Line objects in the history's top and bottom lists,
store LineView of them. This makes them read-only and allows us to make
Line a full dict without an overloaded __missing__.

Skipping __missing__ ensures that the code of Screen does not access an
x position that does not exist by mistake.