Optional use a dummy set for tracking dirty lines
By default Screen tracks which lines are dirty and should be of interest
for the user. This functionality may not be of interest for all the use
cases and this tracking is expensibe, specially for large geometries.

If track_dirty_lines is set to False, the screen.dirty attribute becomes
a dummy or null set that it is always empty saving memory and time.