Make (0, lines-1) the default margin instead of None
Instead of using None as a special case, set the margins to (0, lines-1)
by default.

This may break user code if the user is expecting None as a valid value
or if it is setting it. If required we could make screen.margins a
property and hide the internal implementation.