Make internal _buffer a dict and not a defaultdict
Using a defaultdict can easily introduce false entries in the buffer
making it less sparse. The new Buffer class supports all the dict's operations
but it does not add an entry if the key is missing.

To add a new entry (a new line), do a buffer.line_at(y). This is
equivalent to dict.setdefault(y, new_line()) but avoids the call to
new_line() if an entry y exists.