Impl compressed_display to optionally avoid left/right spaces, top/bottom empty lines
This is an optimization over screen.display where it is possible to
strip left/right spaces of the lines and/or filter top/bottom whole
empty lines.

It is implemented in an opportunistic fashion so it may not fully
strip/filter all that it is expected.

In particular, lines with left or right spaces with non-default
attributes are not stripped; lines that contains only spaces at the top
or bottom are not filtered (even if lstrip/rstrip is set).

This implementation is meant to be used when the screen has very large
geometries and screen.display wastes too much time and memory on padding
large lines and/or filling with a lot of empty lines.