Rework CSI parsing to be aware of subparameters
This fixes https://github.com/selectel/pyte/issues/178.

Before, we assumed that CSI parameters are only integers. However, that
caused us to barf in weird ways when presented with CSI parameters that
contain `:`-delimited subparameters.

This update to the parsing code causes us to parse those subparameters,
and then immediately discard them. That may seem kind of weird, but I'm
laying the groundwork for a followup change to the SGR handling code to
actually be aware of subparameters
(https://github.com/selectel/pyte/issues/179). I just didn't want to
muddy this diff with that change as well.