Refactor CSI param parsing
Now we do the math as we find each digit, rather than at the end after
we have collected each digit.

It's possible this is faster than extending a str and doing an
`int(...)` conversion at the end (I haven't profiled it), but that's not
why I made this change. I made this change to support an upcoming change
to support SGR parameter "context", which is when a SGR CSI parameter
is actually a list of integers: <https://github.com/selectel/pyte/issues/178>.
It's nicer to be able to collect the integers as we go, rather than
having to process them all at the end.